fix: install zlib requirements for brew tests
The brew requirements are moved to the install_homebrew script in order to have them only in one place. For now this is fine since all linux builds have the same requirements. If this changes in the future they can be moved back to the DockerFiles or use an arg or something to configure this.

Additionally, also run extra tests on the Homebrew installation using brew tests. Since this commit (because of installing the requirements), brew audit now also works.